adding pages directory and refactoring
