+++ frontend/src/App.tsx
@@ -3,9 +3,9 @@ import "./App.css";
 
 import { Route, Routes, Outlet, Navigate } from "react-router-dom";
 
-import { UiDemoPageForTest } from "./ui/demo_page";
-import { ObjectsPage } from "./ui/objects_page";
-import { ProjectsPage } from "./ui/projects_page";
+import { UiDemoPageForTest } from "./ui/pages/demo_page";
+import { ObjectsPage } from "./ui/pages/objects_page";
+import { ProjectsPage } from "./ui/pages/projects_page";
 
 import NavBar from "./ui/navbar";
